added the burning ship set, finishing norminette, gave up on zooming to the mouse position

+++ complex_math.c
@@ -33,8 +33,7 @@ t_complex	sqr(t_complex a)
 t_complex	mappoint(t_fractal *fractal, double x, double y)
 {
 	t_complex	c;
-	int			l;
-
+	double		l;
 
 	if (fractal->width < fractal->height)
 		l = fractal->height * fractal->zoom;
